requirements.txt
streamlit
python-dotenv
requests
python-docx
fpdf2
